extract hostname from SCCRQ and add it to nfq output message
HOSTNAME is a mandatory AVP in SCCRQ messages so it should always be
present. Modify the AVP parser to continue walking the AVP list until
both TUNNEL-ID and HOSTNAME AVPs are found.

All L2TP messages contain a MESSAGE_TYPE AVP as the first AVP. If a
peer sends a bad message without this AVP, ensure that we do not
process it.